move extractdata to value.go
