+++ pkg/values/values.go
@@ -2,6 +2,7 @@ package values
 
 import (
 	"context"
+	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/microerror"
@@ -61,6 +62,31 @@ func (v *Values) MergeAll(ctx context.Context, app v1alpha1.App, appCatalog v1al
 	}
 
 	return configMapData, nil
+}
+
+func extractData(resourceType, name string, data map[string]string) (map[string]interface{}, error) {
+	var err error
+	var rawMapData map[string]interface{}
+
+	if data == nil {
+		return rawMapData, nil
+	}
+
+	if len(data) != 1 {
+		return nil, microerror.Maskf(parsingError, "expected %#q %s has only one key but got %d", name, resourceType, len(data))
+	}
+
+	var rawData []byte
+	for _, v := range data {
+		rawData = []byte(v)
+	}
+
+	err = yaml.Unmarshal(rawData, &rawMapData)
+	if err != nil {
+		return nil, microerror.Maskf(parsingError, "failed to parse %#q %s, logs: %s", name, resourceType, err.Error())
+	}
+
+	return rawMapData, nil
 }
 
 // toStringMap converts from a byte slice map to a string map.
